backend/easychef/requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
gunicorn==20.1.0